backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==23.5
reportlab==4.0.9
Pillow==10.1.0
drf-spectacular==0.27.0
